@@ -1,21 +1,22 @@
 {
+  "type": "module",
   "scripts": {
     "build": "tsc",
     "test": "vitest --coverage --ui",
     "start": "node dist/cli.js"
   },
   "dependencies": {
-    "picocolors": "^1.0.0",
-    "vitest": "^1.1.0",
+    "picocolors": "^1.1.1",
+    "vitest": "^3.0.1",
     "yargs": "^17.7.2",
-    "zod": "^3.22.4"
+    "zod": "^3.24.1"
   },
   "devDependencies": {
-    "@types/node": "^20.10.5",
-    "@types/yargs": "^17.0.32",
-    "@vitest/coverage-v8": "^1.1.0",
-    "@vitest/ui": "^1.1.0",
-    "typescript": "^5.3.3",
-    "vite": "^5.0.10"
+    "@types/node": "^22.10.7",
+    "@types/yargs": "^17.0.33",
+    "@vitest/coverage-v8": "^3.0.1",
+    "@vitest/ui": "^3.0.1",
+    "typescript": "^5.7.3",
+    "vite": "^6.0.7"
   }
 }
