@@ -2,20 +2,18 @@
   "name": "messenger",
   "version": "0.0.0",
   "private": true,
+  "type": "module",
   "scripts": {
     "deploy": "wrangler deploy",
     "start": "wrangler dev"
   },
   "devDependencies": {
-    "@cloudflare/workers-types": "^4.20230419.0",
-    "@octokit/webhooks-types": "^7.3.1",
-    "typescript": "^5.0.4",
-    "wrangler": "^3.0.0"
+    "@cloudflare/workers-types": "^4.20250109.0",
+    "@octokit/webhooks-types": "^7.6.1",
+    "typescript": "^5.7.3",
+    "wrangler": "^3.103.1"
   },
   "dependencies": {
-    "@octokit/auth-app": "^6.0.3",
-    "@octokit/core": "^5.0.2",
-    "@octokit/webhooks-methods": "^4.0.0",
-    "octokit": "^3.1.2"
+    "octokit": "^4.1.0"
   }
 }
