@@ -1,3 +1,3 @@
-import { main } from "./main";
+import { main } from "./main.js";
 
 main(process.cwd(), process.argv).then((code) => process.exit(code[0]));
